+++ CHANGES.rst
@@ -38,9 +38,8 @@ Changes
 
 Bugfixes
 --------
-- :class:`DropSimilar` and :func:`column_associations` now work with Polars
-  dataframes when PyArrow is not installed. In that case, Pearson's correlation
-  is omitted from the computed associations.
+- :class:`DropSimilar` now works with Polars dataframes when PyArrow is not
+  installed by avoiding the unused Pearson's correlation computation.
   :pr:`2216` by :user:`Shreyansh Goyal <ShreyanshGoyal>`.
 
 - The parallel coordinate plot created by :meth:`ParamSearch.show_results` could